requirements.txt
deepspeed==0.14.3
einops==0.8.0
GitPython==3.1.43
numpy==1.23.1
pandas==2.2.2
Pillow==10.3.0
python_box==7.1.1
PyYAML==6.0.1
scipy==1.13.1
sentencepiece==0.2.0
tensorboardX==2.6.2.2
termcolor==2.4.0
torch==1.13.0+cu117
torchvision==0.14.0+cu117
